pyproject.toml
[project]
name = "harish-langgraph-newversion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grandalf>=0.8",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-tavily>=0.2.12",
    "langchainhub>=0.1.21",
    "langgraph>=1.0.1",
    "langgraph-checkpoint-sqlite>=3.0.0",
    "notebook>=7.4.7",
    "python-dotenv>=1.2.1",
    "streamlit>=1.51.0",
]
